+++ sysdeps/generic/dl-sysdep.c
@@ -298,8 +298,8 @@ _dl_important_hwcaps (const char *platform, size_t platform_len, size_t *sz,
       if (result == NULL)
 	{
 	no_memory:
-	  _dl_signal_error (ENOMEM, NULL, NULL,
-			    N_("cannot create capability list"));
+	  INT(_dl_signal_error) (ENOMEM, NULL, NULL,
+				 N_("cannot create capability list"));
 	}
 
       result[0].str = (char *) result;	/* Does not really matter.  */
